# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * 火車搖晃程度紀錄系統 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * 單筆搖晃紀錄資料
 */
export interface RecordInput {
  /** 緯度 */
  lat: number;
  /** 經度 */
  lng: number;
  /** 時間戳記（ISO 8601 格式） */
  timestamp: string;
  /** X 軸加速度（左右搖晃）RMS 值 */
  x_accel: number;
  /** Z 軸加速度（上下震動）RMS 值 */
  z_accel: number;
  /**
     * 搖晃等級（1 至 5 級）
     * @minimum 1
     * @maximum 5
     */
  shake_level: number;
  /** 綜合搖晃指數（X 軸與 Z 軸加權合成值） */
  shake_index: number;
  /**
     * 行駛速度（km/h）；由感測器直接取得或以 Haversine 公式從座標推算
     * @minimum 0
     */
  speed_kmh: number;
  /** 速度是否為推算值（true=座標 Haversine 推算，false=感測器實測） */
  speed_estimated: boolean;
}

/**
 * 行程上傳資料
 */
export interface TripUpload {
  /** 裝置識別碼（UUID） */
  device_id: string;
  /** 本次行程識別碼（UUID） */
  trip_id: string;
  /** 備註（可選） */
  note?: string;
  /** 行程資料陣列 */
  records: RecordInput[];
}

/**
 * 上傳結果
 */
export interface UploadResult {
  success: boolean;
  trip_id: string;
  /** 成功寫入的紀錄筆數 */
  count: number;
}

/**
 * 行程摘要資訊
 */
export interface Trip {
  trip_id: string;
  device_id: string;
  /** 行程開始時間 */
  start_time: string;
  /** 行程結束時間 */
  end_time: string;
  /**
     * 備註
     * @nullable
     */
  note?: string | null;
  /** 紀錄筆數 */
  record_count: number;
}

/**
 * 單筆搖晃紀錄（從資料庫取出）
 */
export interface ShakeRecord {
  id: number;
  trip_id: string;
  lat: number;
  lng: number;
  timestamp: string;
  x_accel: number;
  z_accel: number;
  shake_level: number;
  shake_index: number;
  /** 行駛速度（km/h） */
  speed_kmh: number;
  /** 速度是否為推算值 */
  speed_estimated: boolean;
}

/**
 * 行程詳細資料（含所有紀錄點）
 */
export interface TripDetail {
  trip: Trip;
  records: ShakeRecord[];
}

/**
 * 行程備註更新請求
 */
export interface UpdateNoteBody {
  /**
     * 備註文字；傳入 null 可清除備註
     * @nullable
     */
  note: string | null;
}

/**
 * 備註更新結果
 */
export interface UpdateNoteResult {
  success: boolean;
}

/**
 * 累積分析請求
 */
export interface CumulativeAnalysisRequest {
  /**
     * 納入分析的行程識別碼清單（至少一筆）
     * @minItems 1
     */
  trip_ids: string[];
}

/**
 * 單一網格的統計結果
 */
export interface GridCell {
  /** 網格代表座標緯度（網格內所有紀錄的平均緯度） */
  lat: number;
  /** 網格代表座標經度 */
  lng: number;
  /** 代表性搖晃指數（已排除離群值，或樣本不足時的直接平均值） */
  shake_index: number;
  /** 網格內累積的紀錄筆數 */
  count: number;
  /** 是否為樣本數不足網格（true=樣本不足僅供參考，false=統計結果可信） */
  insufficient_samples: boolean;
}

/**
 * 累積分析結果
 */
export interface CumulativeAnalysisResult {
  /** 所有網格的統計結果清單 */
  grids: GridCell[];
}

export interface ErrorResponse {
  error: string;
}