# lib/api-zod/src/generated/types/cumulativeAnalysisRequest.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * 火車搖晃程度紀錄系統 API
 * OpenAPI spec version: 0.1.0
 */

/**
 * 累積分析請求
 */
export interface CumulativeAnalysisRequest {
  /**
     * 納入分析的行程識別碼清單（至少一筆）
     * @minItems 1
     */
  trip_ids: string[];
}