# lib/api-zod/src/generated/types/cumulativeAnalysisResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * 火車搖晃程度紀錄系統 API
 * OpenAPI spec version: 0.1.0
 */
import type { GridCell } from './gridCell';

/**
 * 累積分析結果
 */
export interface CumulativeAnalysisResult {
  /** 所有網格的統計結果清單 */
  grids: GridCell[];
}