# lib/api-zod/src/generated/types/gridCell.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * 火車搖晃程度紀錄系統 API
 * OpenAPI spec version: 0.1.0
 */

/**
 * 單一網格的統計結果
 */
export interface GridCell {
  /** 網格代表座標緯度（網格內所有紀錄的平均緯度） */
  lat: number;
  /** 網格代表座標經度 */
  lng: number;
  /** 代表性搖晃指數（已排除離群值，或樣本不足時的直接平均值） */
  shake_index: number;
  /** 網格內累積的紀錄筆數 */
  count: number;
  /** 是否為樣本數不足網格（true=樣本不足僅供參考，false=統計結果可信） */
  insufficient_samples: boolean;
}