# lib/api-zod/src/generated/types/updateNoteBody.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * 火車搖晃程度紀錄系統 API
 * OpenAPI spec version: 0.1.0
 */

/**
 * 行程備註更新請求
 */
export interface UpdateNoteBody {
  /**
     * 備註文字；傳入 null 可清除備註
     * @nullable
     */
  note: string | null;
}