# lib/api-zod/src/generated/types/updateNoteResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * 火車搖晃程度紀錄系統 API
 * OpenAPI spec version: 0.1.0
 */

/**
 * 備註更新結果
 */
export interface UpdateNoteResult {
  success: boolean;
}